PDF Question Processor/pythonanywhere_deploy/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
requests==2.31.0
google-generativeai==0.3.2
supabase==1.2.0
python-multipart==0.0.6
PyPDF2==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
storage3==0.7.0
postgrest==0.12.0
gotrue==1.3.0
realtime==1.0.0
supafunc==0.3.0
